feat(molecular/Grid): test coverage, split Gird and GridItem (#79)
* feat(molecular/Grid): test coverage, split Gird and GridItem

* release(molecular): v0.1.4-alpha.2

---------
