+++ configs/rollup/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@99mini/rollup-config",
-  "version": "0.1.0",
+  "version": "0.2.0",
   "main": "./src/index.js",
   "devDependencies": {
     "@babel/core": "^7.24.6",
