+++ packages/atom/src/Badge/Badge.css
@@ -1,4 +1,4 @@
-.YnI-Badge-root {
+.YnI-Badge-Root {
   position: relative;
   display: inline-flex;
   width: fit-content;
