+++ packages/core/src/Ripple/Ripple.css
@@ -1,4 +1,4 @@
-.YnI-Ripple-root {
+.YnI-Ripple-Root {
   @keyframes ripple-animation {
     0% {
       transform: translate(-50%, -50%) scale(1);
