+++ packages/molecular/src/Grid/Grid.test.tsx
@@ -1,7 +1,8 @@
 import { render } from "@testing-library/react";
 
 import React from "react";
-import Grid, { GridItem } from "./Grid";
+import Grid from "./Grid";
+import GridItem from "../GridItem";
 
 describe("Grid", () => {
   it("should render without crashing", () => {
