+++ packages/molecular/src/GridItem/GridItem.css
@@ -1,0 +1,7 @@
+.YnI-Grid-Root .YnI-Grid .YnI-Grid-Item {
+  transition: transform 200ms;
+}
+
+.YnI-Grid-Item-Irregular {
+  position: absolute;
+}
