+++ packages/molecular/src/GridItem/index.ts
@@ -1,0 +1,2 @@
+export * from "./GridItem";
+export { default } from "./GridItem";
