+++ packages/molecular/src/TabContext/index.ts
@@ -1,0 +1,7 @@
+export {
+  TabContext,
+  TabContextGenerator,
+  incrementTabId,
+  tabId,
+  type TabContextType,
+} from "./TabContext";
