+++ packages/molecular/src/Toast/Toast.css
@@ -1,4 +1,4 @@
-.YnI-Toast-root {
+.YnI-Toast-Root {
   @keyframes toast-in {
     0% {
       transform: translateY(-1%);
