+++ packages/molecular/src/index.ts
@@ -1,4 +1,5 @@
 export * from "./Grid";
+export * from "./GridItem";
 export * from "./Tab";
 export * from "./Tabs";
 export * from "./Toast";
